requirements.txt
groq
sentence-transformers
rank-bm25
numpy
python-dotenv